refactor: organize binary crate modules
