+++ src/batch/tests.rs
