+++ src/cli/contract_tests.rs
